backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg2-binary==2.9.9